Refactor LoginEvent handling and move to Events namespace
